Add files via upload
The objective of this project is to analyze the historical stock data of Netflix spanning the last 20 years. Through comprehensive analysis, we aim to achieve the following objectives:
1.)Volume of stock traded
2.)Netflix stock price - High, Open, Close
3.)Netflix stock price - Day, Month, Year wise
4.)Top 5 date with Highest Stock Price
5.)Top 5 date with Lowest Stock Price

Now coming to the dataset which is used in this analysis comprises daily or monthly Netflix stock prices over the past two decades. It includes variables such as date, opening price, closing price, high and low prices for the day or month, trading volume, and possibly additional macroeconomic indicators for contextual analysis.